update away from no-namespace errors (avoid namespaces)
