+++ saga/src/types/accounts-client.ts
@@ -1,0 +1,5 @@
+import * as Commands from './commands';
+
+export interface Client {
+  createAccount(params: Commands.CreateAccount): Promise<void>;
+}
