+++ saga/src/types/clients-client.ts
@@ -1,0 +1,6 @@
+import * as Commands from './commands';
+
+export interface Client {
+  addClient(params: Commands.AddClient): Promise<void>;
+  removeClient(params: Commands.RemoveClient): Promise<void>;
+}
