+++ saga/src/types/post-office-client.ts
@@ -1,0 +1,8 @@
+import * as Commands from './commands';
+import AddPostalAddress = Commands.AddPostalAddress;
+import ClearPostalAddresses = Commands.ClearPostalAddresses;
+
+export interface Client {
+  addAddress(params: AddPostalAddress): Promise<void>;
+  clearPostalAddresses(params: ClearPostalAddresses): Promise<void>;
+}
