--- saga/src/types/services.d.ts
@@ -1,3 +1,0 @@
-export interface DB {
-  save(): Promise<string>;
-}
